+++ tinted-builder-rust/src/lib.rs
@@ -1,7 +1,8 @@
 // With the new split of tinted-builder-rust (cli) and
-// tinted-builder (lib), tinted-builder is exported here for bakward
+// tinted-builder (lib), tinted-builder is exported here for backward
 // compatibility for a time. Everyone should move to using
 // tinted-builder as the rust library.
+#[doc = include_str!("../README.md")]
 
 mod operations {
     pub mod build;
@@ -10,8 +11,7 @@ mod utils;
 
 pub use crate::operations::build as operation_build;
 
-/// # Add tinted-builder library test since tinted-builder-rust is
-/// exporting the structs
+/// Add tinted-builder library test since tinted-builder-rust is exporting the structs
 ///
 /// ```
 /// use tinted_builder_rust::{Scheme, Template};
@@ -40,15 +40,14 @@ pub use crate::operations::build as operation_build;
 ///   base0D: "6a9eb5"
 ///   base0E: "78a38f"
 ///   base0F: "a3a079""#;
-/// let template = Template::new(template).unwrap();
-/// let scheme: Scheme = serde_yaml::from_str(&scheme_str).unwrap();
+/// let scheme = Scheme::Base16(serde_yaml::from_str(&scheme_str).unwrap());
+/// let template = Template::new(template, scheme);
 /// let output = template
-///     .render(&scheme)
+///     .render()
 ///     .unwrap();
 ///
 ///  assert_eq!(output, r#"/* Some CSS file with UwUnicorn theme */
 /// .someCssSelector { background-color: #241b26 }
 /// .someOtherCssSelector { background-color: #a3a079 }"#);
 /// ```
-pub use tinted_builder::Scheme;
-pub use tinted_builder::Template;
+pub use tinted_builder::{Base16Scheme, Scheme, Template, TintedBuilderError};