+++ tinted-builder/src/template.rs
@@ -1,112 +1,89 @@
-use anyhow::{Context, Result};
-use std::collections::HashMap;
-use std::fs::{remove_file, File};
-use std::io::Write;
-use std::path::Path;
+mod base16;
 
-use crate::Scheme;
+use crate::{error::TintedBuilderError, scheme::Scheme};
 
+/// A struct representing a template that can be rendered with the provided color scheme.
+///
+/// The `Template` struct holds the content of the template and the scheme used to render it. It
+/// provides methods to create a new template and render it into a `String` using the specified
+/// color scheme.
 pub struct Template {
     content: String,
-}
-
-pub(crate) fn write_to_file(path: &Path, contents: &str) -> Result<()> {
-    if path.exists() {
-        remove_file(path).with_context(|| format!("Unable to remove file: {}", path.display()))?;
-    }
-
-    let mut file =
-        File::create(path).with_context(|| format!("Unable to create file: {}", path.display()))?;
-
-    file.write_all(contents.as_bytes())?;
-
-    Ok(())
+    scheme: Scheme,
 }
 
 impl Template {
-    pub fn new(content: String) -> Result<Template> {
-        Ok(Template { content })
+    /// Creates a new `Template` instance.
+    ///
+    /// # Arguments
+    ///
+    /// * `content` - A `String` representing the content of the template.
+    /// * `scheme` - A `Scheme` enum that determines which color scheme to use when rendering the template.
+    ///
+    /// # Returns
+    ///
+    /// A new `Template` instance with the provided content and scheme.
+    pub fn new(content: String, scheme: Scheme) -> Template {
+        Template { content, scheme }
     }
 
-    fn to_template_context(scheme: &Scheme) -> HashMap<String, String> {
-        let mut context = HashMap::new();
+    /// Renders the template into a `String` using the provided color scheme.
+    ///
+    /// This method applies the specified `Scheme` to the template content, converting placeholders
+    /// in the content to their corresponding values from the scheme context.
+    ///
+    /// # Errors
+    ///
+    /// Returns a `TintedBuilderError` if the rendering process fails. This could happen if the
+    /// content contains placeholders that cannot be resolved using the scheme context.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use tinted_builder::{Template, Scheme};
+    ///
+    /// let scheme_yaml = r#"
+    /// system: "base16"
+    /// name: "Some name"
+    /// author: "Some author"
+    /// variant: "dark"
+    /// palette:
+    ///   base00: "241b26"
+    ///   base01: "2f2a3f"
+    ///   base02: "46354a"
+    ///   base03: "6c3cb2"
+    ///   base04: "7e5f83"
+    ///   base05: "eed5d9"
+    ///   base06: "d9c2c6"
+    ///   base07: "e4ccd0"
+    ///   base08: "877bb6"
+    ///   base09: "de5b44"
+    ///   base0A: "a84a73"
+    ///   base0B: "c965bf"
+    ///   base0C: "9c5fce"
+    ///   base0D: "6a9eb5"
+    ///   base0E: "78a38f"
+    ///   base0F: "a3a079"
+    /// "#;
+    /// let template = Template::new(
+    ///     r#"{{scheme-system}} scheme name is "{{scheme-name}}" and first color is #{{base00-hex}}"#.to_string(),
+    ///     Scheme::Base16(serde_yaml::from_str(scheme_yaml).unwrap())
+    /// );
+    /// let rendered = template.render().unwrap();
+    ///
+    /// assert_eq!(
+    ///     rendered,
+    ///     r#"base16 scheme name is "Some name" and first color is #241b26"#
+    /// );
+    /// ```
+    pub fn render(&self) -> Result<String, TintedBuilderError> {
+        match self.scheme {
+            Scheme::Base16(ref scheme) | Scheme::Base24(ref scheme) => {
+                let ctx = base16::to_template_context(&scheme.clone());
+                let rendered = base16::render(&self.content, &ctx)?;
 
-        context.insert("scheme-name".to_string(), scheme.name.clone());
-        context.insert("scheme-author".to_string(), scheme.author.clone());
-        context.insert(
-            "scheme-description".to_string(),
-            scheme.description.clone().unwrap_or_default(),
-        );
-        context.insert("scheme-slug".to_string(), scheme.slug.clone());
-        context.insert(
-            "scheme-slug-underscored".to_string(),
-            scheme.slug.replace('-', "_"),
-        );
-        context.insert(
-            "scheme-system".to_string(),
-            scheme.system.as_str().to_string(),
-        );
-        context.insert(
-            "scheme-variant".to_string(),
-            scheme.variant.as_str().to_string(),
-        );
-        context.insert(
-            format!("scheme-is-{}-variant", scheme.variant),
-            "true".to_string(),
-        );
-
-        for (name, color) in scheme.palette.iter() {
-            let hex = color.hex.clone();
-            let rgb = color.rgb;
-
-            context.insert(
-                format!("{}-hex", name),
-                format!("{}{}{}", color.hex.0, color.hex.1, color.hex.2),
-            );
-            context.insert(
-                format!("{}-hex-bgr", name),
-                format!("{}{}{}", color.hex.2, color.hex.1, color.hex.0),
-            );
-            context.insert(format!("{}-hex-r", name), hex.0);
-            context.insert(format!("{}-hex-g", name), hex.1);
-            context.insert(format!("{}-hex-b", name), hex.2);
-            context.insert(format!("{}-rgb-r", name), rgb.0.to_string());
-            context.insert(format!("{}-rgb-g", name), rgb.1.to_string());
-            context.insert(format!("{}-rgb-b", name), rgb.2.to_string());
-            context.insert(
-                format!("{}-dec-r", name),
-                format!("{:.8}", rgb.0 as f64 / 255.),
-            );
-            context.insert(
-                format!("{}-dec-g", name),
-                format!("{:.8}", rgb.1 as f64 / 255.),
-            );
-            context.insert(
-                format!("{}-dec-b", name),
-                format!("{:.8}", rgb.2 as f64 / 255.),
-            );
+                Ok(rendered)
+            }
         }
-
-        context
-    }
-
-    pub fn render(&self, scheme: &Scheme) -> Result<String> {
-        let context = serde_yaml::to_string(&Self::to_template_context(scheme))?;
-        let rendered = ribboncurls::render(&self.content, &context, None)?;
-
-        Ok(rendered)
-    }
-
-    #[deprecated(
-        since = "0.4.0",
-        note = "Please use the `render` method instead and write the output to a file yourself."
-    )]
-    pub fn render_to_file(&self, output_path: &Path, scheme: &Scheme) -> Result<&Self> {
-        let context = serde_yaml::to_string(&Self::to_template_context(scheme))?;
-        let rendered = ribboncurls::render(&self.content, &context, None)?;
-
-        write_to_file(output_path, &rendered)?;
-
-        Ok(self)
     }
 }